floats in input

+++ Cholesky.py
@@ -8,7 +8,7 @@ ordem = len(matriz_coef)
 print(matriz_coef)
 for x in range(alt):
     for y in range(larg):
-        matriz_coef[x,y] = int(input("Insira o elemento da \n Linha: " + str(x+1) + " Coluna: " + str(y+1) + ".\n"))
+        matriz_coef[x,y] = float(input("Insira o elemento da \n Linha: " + str(x+1) + " Coluna: " + str(y+1) + ".\n"))
 print("\n")
 print(matriz_coef)
 print("\n")
